Update program 0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -19,6 +19,6 @@ size_t print_list(const list_t *h)
 		h = h->next;
 		c++;
 	}
-	return (c);
+	return ((size_t)c);
 }
